Generic /embed endpoint
